refactoring index.js (add schedule.js)
